+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [0.1.0-alpha28] - 2024-09-25
 
 - improve: real-time output of `score predict` tty mode
+- improve: better default arguments for `score predict`
 
 ## [0.1.0-alpha27] - 2024-09-24